tests/wpt/tests/css/css-anchor-position/anchor-position-multicol-005.html

<!DOCTYPE html>
<title>Overflow pushing anchors to later fragmentainers than querying element in multicol</title>
<link rel="help" href="https://drafts.csswg.org/css-anchor-1/#propdef-anchor-name">
<link rel="help" href="https://drafts.csswg.org/css-anchor-1/#anchor-pos">
<link rel="help" href="https://drafts.csswg.org/css-anchor-1/#anchor-size">
<link rel="author" href="mailto:kojii@chromium.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/check-layout-th.js"></script>
<script src="support/test-common.js"></script>
<style>
.relpos {
  position: relative;
}
.abspos {
  position: absolute;
}
.columns {
  column-count: 2;
  column-fill: auto;
  column-gap: 10px;
  column-width: 100px;
  width: 210px;
  height: 100px;
}
.spacer {
  height: 10px;
  background: pink;
}
.anchor1 {
  anchor-name: --a1;
  margin-left: 10px;
  width: 40px;
  height: 80px;
  background: orange;
}
.target {
  position: absolute;
  background: lime;
  opacity: 1;
}
.target1 {
  left: anchor(--a1 left);
  top: anchor(--a1 top);
  width: anchor-size(--a1 width);
  height: anchor-size(--a1 height);
}
</style>
<body onload="checkLayoutForAnchorPos('.target')">
  <div class="spacer" style="height: 10px"></div>
  <div class="columns">
    <div class="relpos">
      <div style="height: 50px">
        <div class="spacer" style="height: 110px"></div>
        <div class="relpos" style="height: 50px">
          <div class="anchor1 abspos"></div>
        </div>
      </div>
      <div style="height: 50px">
        <div class="target target1"
             data-expected-width=40 data-expected-height=80></div>
      </div>
    </div>
  </div>
</body>
