LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-position-multicol-005-expected.txt


FAIL .target 1 assert_equals:
<div class="target target1" data-expected-width="40" data-expected-height="80"></div>
width expected 40 but got 0
